Select correct client window in WM’s using virtual roots From bug 2185¹
When manually selecting a window, xprop finds the top-level window containing
the pointer and uses XmuClientWindow to attempt to find a client window having
WM_STATE set.

In a WM using virtual roots (e.g. enlightenment, desks other than first) this
fails beacuse the top-level window is a virtual root, and XmuClientWindow seems
to find the first client having WM_STATE set, which only by coincidence will be
the one the user intended to select (i.e. containing the pointer).

The solution is to descend the window hierarchy at the pointer location to find
a window with WM_STATE set.

1] https://bugs.freedesktop.org/show_bug.cgi?id=2185

Signed-off-by: James Cloos <cloos@jhcloos.com>