Make output with unknown width terminal the same as old code
Auto-detecting the terminal width and eliding icons that are too wide
is a nice new feature, but we shouldn't change the behaviour when that
width is not detectable (e.g., when output to a file). Change the
default terminal width from 80 to 144 + 8 to get back to the previous
output format.

Signed-off-by: Keith Packard <keithp@keithp.com>