Correct icon buffer width computation for truecolor terminals
Fix the test when computing size necessary for true color output to
use 'is_truecolor_term' instead of 'is_utf8_locale'. A non-utf8 locale
with a true color terminal would have under-allocated the
buffer. Also, check for truecolor first to match the output code.

This doesn't fix the fact that this code over-allocates by a factor of
two in this case, but that at least won't cause memory corruption.

Signed-off-by: Keith Packard <keithp@keithp.com>