Draw truecolor icons using dual-color half blocks
The vast majority of monospaced fonts are roughly half as wide as they
are tall; to account for this when displaying icons, `xprop` draws each
pixel twice in the horizontal, which nicely evens out the proportions.

This patch utilizes the Lower Half Block character (▄) to obtain the
same result in a quarter of the space. By setting both a background and
a foreground color, we can draw two pixels per character cell, thereby
halving the apparent height of an icon rather than doubling its width.

Signed-off-by: Eden Bynever <edenbynever@proton.me>
Part-of: <https://gitlab.freedesktop.org/xorg/app/xprop/-/merge_requests/10>